use env variables to populate email value
